Refactoring L2CAP implementation
