+++ identity/src/main/java/com/android/identity/GattServer.java
@@ -26,8 +26,6 @@ import android.bluetooth.BluetoothGattServerCallback;
 import android.bluetooth.BluetoothGattService;
 import android.bluetooth.BluetoothManager;
 import android.bluetooth.BluetoothProfile;
-import android.bluetooth.BluetoothServerSocket;
-import android.bluetooth.BluetoothSocket;
 import android.content.Context;
 import android.os.Build;
 import android.util.Log;
@@ -38,11 +36,7 @@ import androidx.annotation.Nullable;
 import com.android.identity.Constants.LoggingFlag;
 
 import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.OutputStream;
-import java.nio.ByteBuffer;
 import java.util.ArrayDeque;
-import java.util.Arrays;
 import java.util.Queue;
 import java.util.UUID;
 
@@ -69,14 +63,13 @@ class GattServer extends BluetoothGattServerCallback {
     BluetoothGattCharacteristic mCharacteristicServer2Client;
     BluetoothGattCharacteristic mCharacteristicIdent;
     BluetoothGattCharacteristic mCharacteristicL2CAP;
-    private BluetoothServerSocket mServerSocket;
     ByteArrayOutputStream mIncomingMessage = new ByteArrayOutputStream();
     Queue<byte[]> mWritingQueue = new ArrayDeque<>();
     boolean writeIsOutstanding = false;
     private BluetoothGattServer mGattServer;
     private BluetoothDevice mCurrentConnection;
     private int mNegotiatedMtu = 0;
-    private BluetoothSocket mSocket;
+    private L2CAPServer mL2CAPServer;
     private byte[] mIdentValue;
     private boolean mUsingL2CAP = false;
 
@@ -187,21 +180,40 @@ class GattServer extends BluetoothGattServerCallback {
         }
 
         if (mUsingL2CAP) {
-            try {
-                mServerSocket = mBluetoothManager.getAdapter().listenUsingInsecureL2capChannel();
-            } catch (IOException e) {
-                // It is unable to listen L2CAP channel
-                Log.w(TAG, "Unable to listen L2CAP channel " + e.getMessage());
-                mServerSocket = null;
-                mUsingL2CAP = false;
-            }
+            // Start L2CAP socket server
+            mL2CAPServer = new L2CAPServer(new L2CAPServer.Listener() {
+                @Override
+                public void onPeerConnected() {
+                    reportPeerConnected();
+                }
+
+                @Override
+                public void onPeerDisconnected() {
+                    reportPeerDisconnected();
+                }
+
+                @Override
+                public void onMessageReceived(@NonNull byte[] data) {
+                    reportMessageReceived(data);
+                }
+
+                @Override
+                public void onError(@NonNull Throwable error) {
+                    reportError(error);
+                }
+            }, mLog.getLoggingFlags());
+
+            // Set using L2CAP to false if it was not able to start the socket server
+            mUsingL2CAP = mL2CAPServer.start(mBluetoothManager.getAdapter());
         }
         return true;
     }
 
     void stop() {
         mInhibitCallbacks = true;
-        closeServerSocket();
+        if (mL2CAPServer != null) {
+            mL2CAPServer.stop();
+        }
         if (mGattServer != null) {
             try {
                 if (mCurrentConnection != null) {
@@ -258,178 +270,24 @@ class GattServer extends BluetoothGattServerCallback {
                 reportError(new Error("Unexpected read request for L2CAP characteristic, not supported"));
                 return;
             }
-            if (mServerSocket == null) {
-                reportError(new Error("Socket for L2CAP not available"));
+            if (mL2CAPServer == null) {
+                reportError(new Error("L2CAP Server not available"));
                 return;
             }
-            int psm = mServerSocket.getPsm();
-            byte[] psmValue = ByteBuffer.allocate(4).putInt(psm).array();
-            if (mLog.isTransportEnabled()) {
-                mLog.transport("PSM Value generated: " + psm + " byte: " + Util.toHex(psmValue));
+            byte[] psmValue = mL2CAPServer.getPsmValue();
+            if (psmValue != null) {
+                mGattServer.sendResponse(device,
+                        requestId,
+                        BluetoothGatt.GATT_SUCCESS,
+                        0,
+                        psmValue);
+                mL2CAPServer.acceptConnection();
+            } else {
+                mUsingL2CAP = false;
             }
-            mGattServer.sendResponse(device,
-                    requestId,
-                    BluetoothGatt.GATT_SUCCESS,
-                    0,
-                    psmValue);
-            acceptConnection();
         } else {
             reportError(new Error("Read on unexpected characteristic with UUID "
                     + characteristic.getUuid()));
-        }
-    }
-
-    private void acceptConnection() {
-        Thread socketServerThread = new Thread(() -> {
-            try {
-                mSocket = mServerSocket.accept();
-                if (mSocket.isConnected()) {
-                    // Stop accepting new connections
-                    mServerSocket.close();
-                    mServerSocket = null;
-                    reportPeerConnected();
-                    listenSocket();
-                }
-            } catch (IOException e) {
-                Log.e(TAG, "Error getting connection from socket server for L2CAP " + e.getMessage(), e);
-                reportError(new Error("Error getting connection from socket server for L2CAP", e));
-            }
-        });
-        socketServerThread.start();
-    }
-
-    private void listenSocket() {
-        if (mLog.isTransportEnabled()) {
-            mLog.transport("Start reading socket input");
-        }
-        // We can check better/right buffer size to use with L2CAP
-        byte[] mmBuffer = new byte[1024 * 16];
-        int numBytes; // bytes returned from read()
-        // Keep listening to the InputStream until an exception occurs.
-        while (mSocket.isConnected()) {
-            try {
-                // Read from the InputStream.
-                numBytes = mSocket.getInputStream().read(mmBuffer, 0, mmBuffer.length);
-                if (numBytes == -1) {
-                    if (mLog.isTransportEnabled()) {
-                        mLog.transport("Stop listening from socket");
-                    }
-                    reportPeerDisconnected();
-                    break;
-                }
-                if (mLog.isTransportEnabled()) {
-                    mLog.transport("Chunk received by socket: (" + numBytes + ") " + Util.toHex(mmBuffer));
-                }
-                // Report message received.
-                mIncomingMessage.write(mmBuffer, 0, numBytes);
-                byte[] entireMessage = mIncomingMessage.toByteArray();
-                long size = getMessageSize(entireMessage);
-                // Check last chunk received
-                // - if bytes received is less than buffer
-                // - or message length is equal as expected size of the message
-                if (numBytes < mmBuffer.length || (size != -1 && entireMessage.length == size)) {
-                    if (mLog.isTransportEnabled()) {
-                        mLog.transport("Data size from message: (" + size + ") message size: (" + entireMessage.length + ")");
-                        mLog.transport("Message: " + Util.toHex(entireMessage));
-                    }
-                    mIncomingMessage.reset();
-                    reportMessageReceived(entireMessage);
-                }
-            } catch (IOException e) {
-                reportError(new Error("Error on listening input stream from socket L2CAP"));
-                break;
-            }
-        }
-    }
-
-    // Returns the message size based in the expected CBOR structure {"data": h'...'}
-    // Works with a map of one element as 'data' key, needs improvement to be flexible
-    private long getMessageSize(byte[] message) {
-        // Check if it is a map with "data" as key
-        if (message == null || message.length < 7) {
-            return -1;
-        }
-        //A1             # map(1)
-        //   64          # text(4)
-        //      64617461 # "data"
-        byte[] mapDataKey = new byte[]{
-                (byte) 0xa1, (byte) 0x64, (byte) 0x64, (byte) 0x61, (byte) 0x74, (byte) 0x61
-        };
-        byte[] messageStart = Arrays.copyOf(message, 6);
-        if (!Arrays.equals(mapDataKey, messageStart)) {
-            return -1;
-        }
-        // Get data byte array
-        byte[] messageData = Arrays.copyOfRange(message, 6, message.length);
-
-        // Get length expected based on the major type as byte array
-        long size = getExpectedValueSize(messageData);
-        if (size < 0) {
-            return -1;
-        }
-        // Plus the messageStart bytes: map, major type and key value length
-        return size + mapDataKey.length;
-    }
-
-    // Returns the length expected plus the length of the bytes used to inform the length in CBOR
-    private long getExpectedValueSize(byte[] data) {
-        if (data == null || data.length < 1) {
-            return -1;
-        }
-        int initialByte = data[0];
-        switch (initialByte & 31) {
-            case 24:
-                // 1 byte length
-                if (data.length < 2) {
-                    return -1;
-                }
-                return data[1] + 2;
-            case 25:
-                // 2 byte length
-                if (data.length < 3) {
-                    return -1;
-                }
-                long value2bytes = 0;
-                value2bytes |= (data[1] & 0xFF) << 8;
-                value2bytes |= (data[2] & 0xFF) << 0;
-                return value2bytes + 3;
-            case 26:
-                // 4 byte length
-                if (data.length < 5) {
-                    return -1;
-                }
-                long value4bytes = 0;
-                value4bytes |= (long) (data[1] & 0xFF) << 24;
-                value4bytes |= (long) (data[2] & 0xFF) << 16;
-                value4bytes |= (long) (data[3] & 0xFF) << 8;
-                value4bytes |= (long) (data[4] & 0xFF) << 0;
-                return value4bytes + 5;
-            case 27:
-                // 8 byte length
-                if (data.length < 9) {
-                    return -1;
-                }
-                long value8bytes = 0;
-                value8bytes |= (long) (data[1] & 0xFF) << 56;
-                value8bytes |= (long) (data[2] & 0xFF) << 48;
-                value8bytes |= (long) (data[3] & 0xFF) << 40;
-                value8bytes |= (long) (data[4] & 0xFF) << 32;
-                value8bytes |= (long) (data[5] & 0xFF) << 24;
-                value8bytes |= (long) (data[6] & 0xFF) << 16;
-                value8bytes |= (long) (data[7] & 0xFF) << 8;
-                value8bytes |= (long) (data[8] & 0xFF) << 0;
-                return value8bytes + 9;
-            case 28:
-            case 29:
-            case 30:
-                // Reserved 28-30
-                return -1;
-            case 31:
-                // Indefinite
-                return -1;
-            default:
-                // Direct 0-23
-                return (initialByte & 31) + 1;
         }
     }
 
@@ -447,7 +305,7 @@ class GattServer extends BluetoothGattServerCallback {
         if (charUuid.equals(mCharacteristicStateUuid) && value.length == 1) {
             if (value[0] == 0x01) {
                 // Close server socket when the connection was done by state characteristic
-                closeServerSocket();
+                stopL2CAPServer();
                 reportPeerConnected();
             } else if (value[0] == 0x02) {
                 reportTransportSpecificSessionTermination();
@@ -485,25 +343,11 @@ class GattServer extends BluetoothGattServerCallback {
         }
     }
 
-    private void closeServerSocket() {
-        try {
-            if (mServerSocket != null) {
-                mServerSocket.close();
-                mServerSocket = null;
-            }
-        } catch (IOException e) {
-            // Ignoring this error
-            Log.e(TAG, " Error closing server socket connection " + e.getMessage(), e);
-        }
-        try {
-            if (mSocket != null && mSocket.isConnected()) {
-                // Stop accepting new connections
-                mSocket.close();
-                mSocket = null;
-            }
-        } catch (IOException e) {
-            // Ignoring this error
-            Log.e(TAG, " Error closing server socket connection " + e.getMessage(), e);
+    private void stopL2CAPServer() {
+        // Close server socket when the connection was done by state characteristic
+        if (mL2CAPServer != null) {
+            mL2CAPServer.stop();
+            mL2CAPServer = null;
         }
         mUsingL2CAP = false;
     }
@@ -592,18 +436,8 @@ class GattServer extends BluetoothGattServerCallback {
         }
 
         // Uses socket L2CAP when it is available
-        if (mSocket != null && mSocket.isConnected()) {
-            if (mLog.isTransportVerboseEnabled()) {
-                mLog.transportVerbose("sendMessage using L2CAP socket");
-            }
-            try {
-                final OutputStream os = mSocket.getOutputStream();
-                os.write(data);
-                os.flush();
-            } catch (IOException e) {
-                Log.e(TAG, "Error sending message by L2CAP socket " + e.getMessage(), e);
-                reportError(new Error("Error sending message by L2CAP socket"));
-            }
+        if (mL2CAPServer != null && mL2CAPServer.isConnected()) {
+            mL2CAPServer.sendMessage(data);
             return;
         }
